Restructure packaging for molpay
COAND-497
