+++ molpay/src/main/java/com/adyen/checkout/molpay/MolpayComponent.kt
@@ -16,6 +16,7 @@ import com.adyen.checkout.components.util.PaymentMethodTypes
 import com.adyen.checkout.issuerlist.IssuerListComponent
 import com.adyen.checkout.issuerlist.internal.ui.IssuerListDelegate
 import com.adyen.checkout.molpay.MolpayComponent.Companion.PROVIDER
+import com.adyen.checkout.molpay.internal.provider.MolpayComponentProvider
 
 /**
  * Component should not be instantiated directly. Instead use the [PROVIDER] object.
